+++ R/footprinting.R
@@ -27,6 +27,9 @@ GetFootprintData <- function(
   idents = NULL
 ) {
   assay <- SetIfNull(x = assay, y = DefaultAssay(object = object))
+  if (!inherits(x = object[[assay]], what = "ChromatinAssay")) {
+    stop("The requested assay is not a ChromatinAssay")
+  }
   positionEnrichment <- GetAssayData(
     object = object,
     assay = assay,