+++ NEWS.md
@@ -12,6 +12,8 @@ Bug fixes:
 * Fixed fragment file reading when only one fragment found in requested region ([#474](https://github.com/timoast/signac/issues/474))
 * Fixed `standard.chromosomes` parameter in `GetGRangesFromEnsDb()` ([#513](https://github.com/timoast/signac/issues/513))
 * Fixed `group.by` parameter in `PlotFootprint()` ([#522](https://github.com/timoast/signac/issues/522))
+* Fixed bug that would cause some gene coordinates used by `GeneActivity()` to be 
+incorrect ([#521](https://github.com/timoast/signac/issues/521))
 
 Other changes: