+++ NEWS.md
@@ -23,6 +23,7 @@ Other changes:
 on non-`ChromatinAssay` assays
 * Added `biotypes` parameter to `GeneActivity()` and `GetTSSPositions()` functions
 * Performance improvements
+* Changed default value for `n` in `NucleosomeSignal()`
 
 # Signac 1.1.1