+++ NEWS.md
@@ -7,6 +7,7 @@ New functionality:
 include bigWig files in `CoveragePlot()`
 * Added `region.highlight` parameter to `CoveragePlot()`
 * Added `biotypes` parameter to `GeneActivity()` and `GetTSSPositions()` functions
+
 Bug fixes:
 
 * Fixed fragment file reading when only one fragment found in requested region ([#474](https://github.com/timoast/signac/issues/474))